# src/api_help/openapi.json
{
   "openapi": "3.1.0",
   "info": {
      "title": "Concierge",
      "description": "An API to access the Concierge system",
      "version": "1.0.0"
   },
   "servers": [
      {
         "url": "https://example.com/",
         "description": "Production server"
      }
   ],
   "components": {
      "schemas": {
         "Task": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "integer",
                  "description": "task's id",
                  "example": "123"
               },
               "taskType": {
                  "type": "string",
                  "description": "Type of the tasks",
                  "example": "Maintenance"
               },
               "description": {
                  "type": "string",
                  "description": "description of tasks",
                  "example": "There is a leak in the bathroom sink that needs urgent attention."
               },
               "roomId": {
                  "type": "integer",
                  "description": "Location of the task. Primarily used for room-related tasks",
                  "example": "123"
               },
               "requesterId": {
                  "type": "integer",
                  "description": "Account ID of requester",
                  "example": "10"
               },
               "assigneeId": {
                  "type": "integer",
                  "description": "Account ID of responsible staff",
                  "example": "12"
               },
               "status": {
                  "type": "string",
                  "description": "Indicates if a task is complete or not",
                  "example": "InProgress"
               },
               "createdAt": {
                  "type": "string",
                  "description": "Date and time this task was created",
                  "example": "2024-10-17T10:00:00"
               }
            }
         },
         "AddTaskDto": {
            "required": [
               "description",
               "requesterId",
               "roomId",
               "taskType"
            ],
            "type": "object",
            "properties": {
               "taskType": {
                  "$ref": "#/components/schemas/TaskItemType"
               },
               "description": {
                  "maxLength": 400,
                  "minLength": 1,
                  "type": "string"
               },
               "roomId": {
                  "maximum": 500,
                  "minimum": 100,
                  "type": "integer",
                  "format": "int32"
               },
               "requesterId": {
                  "maximum": 500,
                  "minimum": 1,
                  "type": "integer",
                  "format": "int32"
               }
            },
            "additionalProperties": false
         },
         "TaskItemStatus": {
            "enum": [
               "Pending",
               "InProgress",
               "Completed"
            ],
            "type": "string"
         },
         "TaskItemType": {
            "enum": [
               "RoomCleaning",
               "Maintenance",
               "FoodDelivery",
               "WakeUpCall",
               "LaundryService",
               "SpaAndMassage"
            ],
            "type": "string"
         },
         "UpdateAssigneeDto": {
            "required": [
               "assigneeId"
            ],
            "type": "object",
            "properties": {
               "assigneeId": {
                  "type": "integer",
                  "format": "int32"
               }
            },
            "additionalProperties": false
         },
         "UpdateTaskDto": {
            "required": [
               "description",
               "status",
               "taskType"
            ],
            "type": "object",
            "properties": {
               "taskType": {
                  "$ref": "#/components/schemas/TaskItemType"
               },
               "description": {
                  "maxLength": 400,
                  "minLength": 1,
                  "type": "string"
               },
               "assigneeId": {
                  "type": "integer",
                  "format": "int32",
                  "nullable": true
               },
               "status": {
                  "$ref": "#/components/schemas/TaskItemStatus"
               }
            },
            "additionalProperties": false
         },
...
      }
   },
   "paths": {
...
      "tasks": {
         "summary": "View, add or modify hotel service tasks",
         "get": {
            "summary": "View All Tasks",
            "description": "Requests /permissions/{permission-name-or-id}/sessions/{X-API-KEY} for access permission",
            "operationId": "getTasks",
            "tags": [
               "Tasks"
            ],
            "parameters": [
               {
                  "name": "X-API-KEY",
                  "in": "cookie",
                  "description": "A session key, associated to an account with the necessary permissions for authenticating user",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               },
               {
                  "name": "RoomId",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "RequesterId",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "AssigneeId",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "Status",
                  "in": "query",
                  "schema": {
                     "$ref": "#/components/schemas/TaskItemStatus"
                  }
               },
               {
                  "name": "Year",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "Month",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "Day",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               },
               {
                  "name": "SortAscending",
                  "in": "query",
                  "schema": {
                     "type": "boolean"
                  }
               },
               {
                  "name": "AccountId",
                  "in": "query",
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "tasks": {
                                 "type": "array",
                                 "items": {
                                    "$ref": "#/components/schemas/Task"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "400": {
                  "description": "Bad request"
               },
               "401": {
                  "description": "Unauthorized"
               }
            }
         },
         "post": {
            "summary": "Create Task",
            "description": "Requests GET /sessions/{sessionId} to get accountID associated to current session.",
            "operationId": "postTask",
            "tags": [
               "Tasks"
            ],
            "parameters": [
               {
                  "name": "X-API-KEY",
                  "in": "cookie",
                  "description": "A session key, associated to an account with the necessary permissions for authenticating user",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               }
            ],
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/AddTaskDto"
                     }
                  },
                  "text/json": {
                     "schema": {
                        "$ref": "#/components/schemas/AddTaskDto"
                     }
                  },
                  "application/*+json": {
                     "schema": {
                        "$ref": "#/components/schemas/AddTaskDto"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "OK"
               },
               "400": {
                  "description": "Bad request",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "object",
                           "properties": {
                              "error": {
                                 "type": "string",
                                 "example": "The 'taskType' field is required"
                              }
                           }
                        }
                     }
                  }
               },
               "401": {
                  "description": "Unauthorized"
               }
            }
         }
      },
      "/tasks/{id}": {
         "get": {
            "summary": "View task",
            "operationId": "getTaskById",
            "tags": [
               "Tasks"
            ],
            "parameters": [
               {
                  "name": "X-API-KEY",
                  "in": "cookie",
                  "description": "A session key, associated to an account with the necessary permissions for authenticating user",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               },
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/Task"
                        }
                     }
                  }
               },
               "400": {
                  "description": "Bad request"
               },
               "401": {
                  "description": "Unauthorized"
               }
            }
         },
         "put": {
            "summary": "Update Task",
            "description": "Requests GET /sessions/{sessionId} to get accountID associated to current session.",
            "operationId": "putTask",
            "tags": [
               "Tasks"
            ],
            "parameters": [
               {
                  "name": "X-API-KEY",
                  "in": "cookie",
                  "description": "A session key, associated to an account with the necessary permissions for authenticating user",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               },
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/UpdateTaskDto"
                     }
                  },
                  "text/json": {
                     "schema": {
                        "$ref": "#/components/schemas/UpdateTaskDto"
                     }
                  },
                  "application/*+json": {
                     "schema": {
                        "$ref": "#/components/schemas/UpdateTaskDto"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "OK"
               },
               "400": {
                  "description": "Bad request"
               },
               "401": {
                  "description": "Unauthorized"
               }
            }
         },
         "delete": {
            "summary": "Delete Task",
            "description": "Requests GET /sessions/{sessionId} to get accountID associated to current session.",
            "operationId": "deleteTask",
            "tags": [
               "Tasks"
            ],
            "parameters": [
               {
                  "name": "X-API-KEY",
                  "in": "cookie",
                  "description": "A session key, associated to an account with the necessary permissions for authenticating user",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               },
               {
                  "name": "task-id",
                  "in": "path",
                  "description": "Id of task to delete",
                  "required": true,
                  "schema": {
                     "type": "integer"
                  }
               }
            ],
            "responses": {
               "200": {
                  "description": "OK",
                  "content": {
                     "application/json": {
                        "schema": {
                           "type": "boolean",
                           "example": true
                        }
                     }
                  }
               },
               "400": {
                  "description": "Bad request"
               },
               "401": {
                  "description": "Unauthorized"
               }
            }
         }
      },
      "/tasks/{id}/assignee": {
         "patch": {
            "summary": "Assign Task",
            "description": "Requests /permissions/{permission-name-or-id}/sessions/{session-key} for access permission.",
            "operationId": "patchTask",
            "tags": [
               "Tasks"
            ],
            "parameters": [
               {
                  "name": "X-API-KEY",
                  "in": "cookie",
                  "description": "A session key, associated to an account with the necessary permissions for authenticating user",
                  "required": true,
                  "schema": {
                     "type": "string"
                  }
               },
               {
                  "name": "id",
                  "in": "path",
                  "required": true,
                  "schema": {
                     "type": "integer",
                     "format": "int32"
                  }
               }
            ],
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/UpdateAssigneeDto"
                     }
                  },
                  "text/json": {
                     "schema": {
                        "$ref": "#/components/schemas/UpdateAssigneeDto"
                     }
                  },
                  "application/*+json": {
                     "schema": {
                        "$ref": "#/components/schemas/UpdateAssigneeDto"
                     }
                  }
               }
            },
            "responses": {
               "200": {
                  "description": "OK"
               },
               "400": {
                  "description": "Bad request"
               },
               "401": {
                  "description": "Unauthorized"
               }
            }
         }
      },
...
   }
}